Plot neural network architecture

plotnn It is designed for an inspection of the weights for objects of class statnn.

Usage

plotnn(
  x,
  rep = NULL,
  x.entry = NULL,
  x.out = NULL,
  radius = 0.15,
  arrow.length = 0.2,
  intercept = FALSE,
  intercept.factor = 0.4,
  information = FALSE,
  information.pos = 0.1,
  col.entry.synapse = "black",
  col.entry = "black",
  col.hidden = "black",
  col.hidden.synapse = "black",
  col.out = "black",
  col.out.synapse = "black",
  col.intercept = "black",
  col.sig.synapse = "black",
  col.insig.synapse = "lightgrey",
  col.sig.node = "black",
  col.insig.node = "lightgrey",
  fontsize = 12,
  dimension = 6,
  show.weights = FALSE,
  file = NULL,
  rounding = 3,
  alpha = 0.05,
  lambda = 0,
  ...
)

Arguments

x

an object of class statnnet

rep

repetition of the neural network. If rep="best", the repetition with the smallest error will be plotted. If not stated all repetitions will be plotted, each in a separate window.

x.entry

x-coordinate of the entry layer. Depends on the arrow.length in default.

x.out

x-coordinate of the output layer.

radius

radius of the neurons.

arrow.length

length of the entry and out arrows.

intercept

a logical value indicating whether to plot the intercept.

intercept.factor

x-position factor of the intercept. The closer the factor is to 0, the closer the intercept is to its left neuron.

information

a logical value indicating whether to add the error and steps to the plot.

information.pos

y-position of the information.

col.entry.synapse

color of the synapses leading to the input neurons.

col.entry

color of the input neurons.

col.hidden

color of the neurons in the hidden layer.

col.hidden.synapse

color of the weighted synapses.

col.out

color of the output neurons.

col.out.synapse

color of the synapses leading away from the output neurons.

col.intercept

color of the intercept.

col.sig.synapse

color of the significant synapses.

col.insig.synapse

color of the insignificant synapses.

col.sig.node

color of the significant input nodes.

col.insig.node

color of the insignificant input nodes.

fontsize

fontsize of the text.

dimension

size of the plot in inches.

show.weights

a logical value indicating whether to print the calculated weights above the synapses.

file

a character string naming the plot to write to. If not stated, the plot will not be saved.

rounding

number of decimal places to round values.

alpha

significane level.

lambda

ridge penalty.

...

arguments to be passed to methods, such as graphical parameters (see par).

Value

plot of weights and their significance